+++ examples/main/shared/editor/Editor.tsx
@@ -84,10 +84,6 @@ export default class Editor extends PureComponent<ExampleEditorProps> {
 
   initEditorProps() {
     /* eslint-disable no-console */
-    const onPluginAction: OnPluginAction = async (
-      eventName: EventName,
-      params: PluginEventParams
-    ) => console.log(eventName, params);
     this.helpers = {
       //these are for testing purposes only
       onPluginAdd: async (...args) => console.log('onPluginAdd', ...args),
@@ -116,7 +112,7 @@ export default class Editor extends PureComponent<ExampleEditorProps> {
           updateEntity(testVideo);
         }, mockTimout || 500);
       },
-      onPluginAction,
+      onPluginAction: async (eventName, params) => console.log(eventName, params),
     };
     /* eslint-enable no-console */
     this.setImageUploadHelper();